requirements.txt
django==4.2.3
djangorestframework==3.14.0
django-phonenumber-field==7.1.0
phonenumbers==8.13.17
djangorestframework-simplejwt==5.2.2
model_bakery==1.12.0
pytest-cov==4.1.0
pytest-django==4.5.2
tox==4.6.4